fix app.py

+++ app.py
@@ -34,15 +34,41 @@ app.mount("/static", StaticFiles(directory="static"), name="static")
 
 @app.get('/')
 async def welcome(request:Request) :
+    cookies = request.cookies
+    session_value = cookies.get("session")
+    res = postsSession.find_one({"Session": session_value})
+    if(res==None):
+        return RedirectResponse(url="/auth")
     return templates.TemplateResponse(name='index.html',context={'request':request})
 
 @app.get('/auth')
 async def welcome(request:Request) :
+    cookies = request.cookies
+    session_value = cookies.get("session")
+    res = postsSession.find_one({"Session": session_value})
+    if(res!=None):
+        return RedirectResponse(url="/")
     return templates.TemplateResponse(name='auth.html',context={'request':request})
 
 @app.get('/register')
 async def welcome(request:Request) :
+    cookies = request.cookies
+    session_value = cookies.get("session")
+    res = postsSession.find_one({"Session": session_value})
+    if(res!=None):
+        return RedirectResponse(url="/")
     return templates.TemplateResponse(name='register.html',context={'request':request})
+
+async def new_delete_file(name: str):
+    path = "static/images/"
+    path+=name
+    try:
+        if(path!="static/images/"):
+            os.remove(path)
+    except FileNotFoundError:
+        pass
+    return {"message": "Файл успешно удалён"}
+
 
 @app.put('/data')
 async def upload(request: Request):
@@ -50,28 +76,48 @@ async def upload(request: Request):
     session_value = cookies.get("session")
     res = postsSession.find_one({"Session": session_value})
     if(res==None):
-        return {"status": 403}
-    data = await request.json()
-    email = data["email"]
-    numFilter = data["numFilter"]
-    value0 = data["value0"]
-    value1 = data["value1"]
-    value2 = data["value2"]
-    value3 = data["value3"]
-    value4 = data["value4"]
-    value5 = data["value5"]
-    value6 = data["value6"]
-    filter = {'email':res["id"],'number': numFilter}
+        raise HTTPException(status_code=403, detail="Аккаунт не найден")
 
-    posts.update_many(filter, {'$set': {'type_device': value0}})
-    posts.update_many(filter, {'$set': {'model_device': value1}})
-    posts.update_many(filter, {'$set': {'serial_number': value2}})
-    posts.update_many(filter, {'$set': {'ITAM_device': value3}})
-    posts.update_many(filter, {'$set': {'photo_device': value4}})
-    posts.update_many(filter, {'$set': {'photo_serial_number_device': value5}})
-    posts.update_many(filter, {'$set': {'photo_ITAM_device': value6}})
+    form_data = await request.form()
+    n=posts.count_documents({})
+    newN = posts.find({'user_id': res["id"] })
+    count = int(len(list(newN)))+1
+    id=form_data.get("id")
+    photo_device = form_data.get("photo_device")
+    photo_serial_number_device = form_data.get("photo_serial_number_device")
+    photo_ITAM_device = form_data.get("photo_ITAM_device")
 
-    return {"message": "true"}
+    check = posts.find_one({"user_id":res["id"],"id": int(id)})
+    if(check==None):
+        raise HTTPException(status_code=404, detail="Запись не найдена")
+    filter = {'user_id':res["id"],'id': int(id)}
+
+    posts.update_many(filter, {'$set': {'type_device': form_data.get("type_device")}})
+    posts.update_many(filter, {'$set': {'model_device': form_data.get("model_device")}})
+    posts.update_many(filter, {'$set': {'serial_number': form_data.get("serial_number")}})
+    posts.update_many(filter, {'$set': {'ITAM_device': form_data.get("ITAM_device")}})
+    delete = posts.find_one({'user_id':res["id"],'id': int(id)})
+    if(photo_device!="undefined"):
+        await new_delete_file(delete['photo_device'])
+        name_photo_device = await newUpload(photo_device)
+        posts.update_many(filter, {'$set': {'photo_device': name_photo_device}})
+    if(photo_serial_number_device!="undefined"):
+        await new_delete_file(delete['photo_serial_number_device'])
+        name_photo_serial_number_device = await newUpload(photo_serial_number_device)
+        posts.update_many(filter, {'$set': {'photo_serial_number_device': name_photo_serial_number_device}})
+    if(photo_ITAM_device!="undefined"):
+        await new_delete_file(delete['photo_ITAM_device'])
+        name_photo_ITAM_device = await newUpload(photo_ITAM_device)
+        posts.update_many(filter, {'$set': {'photo_ITAM_device': name_photo_ITAM_device}})
+
+    # delete1 = form_data.get("delete1")
+    # await new_delete_file(delete1)
+    # delete2 = form_data.get("delete2")
+    # await new_delete_file(delete2)
+    # delete3 = form_data.get("delete3")
+    # await new_delete_file(delete3)
+
+    return {"message": "Запись успешно изменена"}
 
 @app.get("/data")
 async def read_data(request: Request):
@@ -79,8 +125,8 @@ async def read_data(request: Request):
     session_value = cookies.get("session")
     res = postsSession.find_one({"Session": session_value})
     if(res==None):
-        return {"status": 403}
-    dataSend = posts.find({'email': res["id"]})
+        raise HTTPException(status_code=403, detail="Аккаунт не найден")
+    dataSend = posts.find({'user_id': res["id"]})
     posts_list = []
     for post in dataSend:
         post_dict = dict(post)
@@ -94,34 +140,32 @@ async def delete(request: Request,numDelete: str):
     session_value = cookies.get("session")
     res = postsSession.find_one({"Session": session_value})
     if(res==None):
-        return {"status": 403}
+        raise HTTPException(status_code=403, detail="Аккаунт не найден")
     count = posts.count_documents({})
-    file = posts.find_one({"email":res["id"],"number": int(numDelete)})
+    file = posts.find_one({"user_id":res["id"],"id": int(numDelete)})
+    if(file==None):
+        raise HTTPException(status_code=404, detail="Запись не найдена")
     path = "static/images/"
     fil1 = path+file["photo_device"]
     fil2 = path+file["photo_serial_number_device"]
     fil3 = path+file["photo_ITAM_device"]
     try:
-        os.remove(fil1)
-        os.remove(fil2)
-        os.remove(fil3)
+        if(fil1!="static/images/"):
+            os.remove(fil1)
+        if(fil2!="static/images/"):
+            os.remove(fil2)
+        if(fil3!="static/images/"):
+            os.remove(fil3)
     except FileNotFoundError:
         pass
-    filterDelete = {'email':res["id"],'number': int(numDelete)}
+    filterDelete = {'user_id':res["id"],'id': int(numDelete)}
     posts.delete_one(filterDelete)
     for i in range(int(numDelete)+1,count+1):
-        filter = {'email':res["id"],'number': i}
-        result = posts.update_one(filter, {'$set': {'number': i-1}})
+        filter = {'user_id':res["id"],'id': i}
+        result = posts.update_one(filter, {'$set': {'id': i-1}})
+    return {"message": "Запись успешно удалена"}
 
-    return {"message": "true"}
-
-@app.post('/uploadFile')
-async def uploadFile(file: UploadFile, request: Request):
-    cookies = request.cookies
-    session_value = cookies.get("session")
-    res = postsSession.find_one({"Session": session_value})
-    if(res==None):
-        return {"status": 403}
+async def newUpload(file: UploadFile):
     newName = "".join([alphabet[random.randint(0, len(alphabet) -1)] for _ in range(0, 60)])
     exp=f".{file.filename.rsplit('.', 1)[1]}"
     newName +=exp
@@ -136,25 +180,31 @@ async def sendForm(request: Request):
     session_value = cookies.get("session")
     res = postsSession.find_one({"Session": session_value})
     if(res==None):
-        return {"status": 403}
+        raise HTTPException(status_code=403, detail="Аккаунт не найден")
     form_data = await request.form()
     n=posts.count_documents({})
-    newN = posts.find({'email': res["id"] })
+    newN = posts.find({'user_id': res["id"] })
     count = int(len(list(newN)))+1
+    photo_device = form_data.get("photo_device")
+    photo_serial_number_device = form_data.get("photo_serial_number_device")
+    photo_ITAM_device = form_data.get("photo_ITAM_device")
+    name_photo_device = await newUpload(photo_device)
+    name_photo_serial_number_device = await newUpload(photo_serial_number_device)
+    name_photo_ITAM_device = await newUpload(photo_ITAM_device)
     if( posts.count_documents({}) == n):
         post = {
-            "email": res["id"],
-            "number": count,
+            "user_id": res["id"],
+            "id": count,
             "type_device": form_data.get("type_device"),
             "model_device": form_data.get("model_device"),
             "serial_number": form_data.get("serial_number"),
             "ITAM_device": form_data.get("ITAM_device"),
-            "photo_device": form_data.get("photo_device"),
-            "photo_serial_number_device": form_data.get("photo_serial_number_device"),
-            "photo_ITAM_device": form_data.get("photo_ITAM_device"),
+            "photo_device": name_photo_device,
+            "photo_serial_number_device": name_photo_serial_number_device,
+            "photo_ITAM_device": name_photo_ITAM_device,
         }
         res = posts.insert_one(post).inserted_id
-    return ""
+    return {"message": "Запись успешно добавлена"}
 
 @app.delete("/file/{name}")
 async def delete_file(request: Request,name: str):
@@ -162,14 +212,15 @@ async def delete_file(request: Request,name: str):
     session_value = cookies.get("session")
     res = postsSession.find_one({"Session": session_value})
     if(res==None):
-        return {"status": 403}
+        raise HTTPException(status_code=403, detail="Аккаунт не найден")
     path = "static/images/"
     path+=name
     try:
-        os.remove(path)
+        if(path!="static/images/"):
+            os.remove(path)
     except FileNotFoundError:
         pass
-    return {"info": f"file {path} deleted"}
+    return {"message": "Файл успешно удалён"}
 
 @app.post("/signIn")
 async def signIn(request: Request):
@@ -178,12 +229,12 @@ async def signIn(request: Request):
     password = data["password"]
     res = postsPassword.find_one({'email': email})
     if(res==None):
-        return {"status" : 403}
+        raise HTTPException(status_code=403, detail="Аккаунт не найден")
     salt = res['salt']
     hash = res['password']
     newHash = hashlib.pbkdf2_hmac('sha256', password.encode('utf-8'), salt, 100000).hex()
     if(hash != newHash):
-        return {"status" : 403}
+        raise HTTPException(status_code=422, detail="Не правильный пароль")
     hashSession = hashlib.pbkdf2_hmac('sha256', email.encode('utf-8'),os.urandom(16).hex().encode('utf-8'), 100000).hex()
     content = {"message": "true"}
     response = JSONResponse(content=content)
@@ -212,9 +263,9 @@ async def signUp(request: Request):
     password = data["password"]
     res = postsPassword.find_one({'email': email})
     if(res!=None):
-        return {"status" : 403}
+        raise HTTPException(status_code=403, detail="Аккаунт не найден")
     if(len(password)<8):
-        return {"status" : 422}
+        raise HTTPException(status_code=422, detail="Пароль слишком короткий")
     n=postsPassword.count_documents({})
     salt = os.urandom(16)
     hashPassword = hashlib.pbkdf2_hmac('sha256', password.encode('utf-8'), salt.hex().encode('utf-8'), 100000)
